:truck: move read modle to `certification\shared` context
