+++ api/src/certification/shared/domain/models/CertificationChallenge.js
