+++ api/tests/tooling/domain-builder/factory/build-certification-challenge-with-type.js
@@ -1,4 +1,4 @@
-import { CertificationChallengeWithType } from '../../../../src/shared/domain/models/CertificationChallengeWithType.js';
+import { CertificationChallengeWithType } from '../../../../src/certification/shared/domain/models/CertificationChallengeWithType.js';
 import { Challenge } from '../../../../src/shared/domain/models/Challenge.js';
 
 const buildCertificationChallengeWithType = function ({
