+++ api/tests/tooling/domain-builder/factory/build-certification-challenge.js
@@ -1,4 +1,4 @@
-import { CertificationChallenge } from '../../../../src/shared/domain/models/CertificationChallenge.js';
+import { CertificationChallenge } from '../../../../src/certification/shared/domain/models/CertificationChallenge.js';
 import { buildSkill } from './build-skill.js';
 
 const buildCertificationChallenge = function ({
